Change `HelpRepository` to use the `TextStore` instead of `FileStorage`
