+++ app/src/androidTest/java/org/simple/clinic/help/HelpSyncAndroidTest.kt
@@ -7,10 +7,10 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.simple.clinic.TestClinicApp
+import org.simple.clinic.help.HelpRepository.Companion.HELP_KEY
 import org.simple.clinic.rules.LocalAuthenticationRule
+import org.simple.clinic.storage.text.TextStore
 import org.simple.clinic.util.Rules
-import org.simple.clinic.util.toNullable
-import org.simple.clinic.util.unwrapJust
 import javax.inject.Inject
 
 
@@ -21,6 +21,9 @@ class HelpSyncAndroidTest {
 
   @Inject
   lateinit var helpSync: HelpSync
+
+  @Inject
+  lateinit var textStore: TextStore
 
   @get:Rule
   val rules: RuleChain = Rules
@@ -34,30 +37,23 @@ class HelpSyncAndroidTest {
 
   @After
   fun tearDown() {
-    val (helpFile) = helpRepository.helpFile().blockingFirst()
-    helpFile?.delete()
+    textStore.delete(HELP_KEY)
   }
 
   @Test
   fun when_pulling_help_from_the_server_it_should_save_the_content_as_a_file() {
-    // Sometimes this file is present on the device when running the test which causes the test
-    // to fail. This is to ensure the file is not present.
-    helpRepository
-        .helpFile()
-        .blockingFirst()
-        .toNullable()
-        ?.delete()
+    textStore.delete(HELP_KEY)
+    // Technically, it should be `null`. But due to an implementation
+    // detail, `AndroidFileStorage` will end up creating an empty file
+    // when `get()` is called, so it will never be `null`.
+    // This will be fixed when we switch to a Room DAO.
+    assertThat(textStore.get(HELP_KEY)).isEmpty()
 
     helpSync
         .pull()
         .test()
         .assertNoErrors()
 
-    val helpFile = helpRepository
-        .helpFile()
-        .unwrapJust()
-        .blockingFirst()
-
-    assertThat(helpFile.length()).isGreaterThan(0L)
+    assertThat(textStore.get(HELP_KEY)).isNotEmpty()
   }
 }
