+++ app/src/main/java/org/simple/clinic/storage/StorageModule.kt
@@ -9,17 +9,15 @@ import dagger.Provides
 import org.simple.clinic.AppDatabase
 import org.simple.clinic.di.AppScope
 import org.simple.clinic.di.AppSqliteOpenHelperFactory
-import org.simple.clinic.storage.files.FileStorage
 import org.simple.clinic.storage.migrations.RoomMigrationsModule
-import org.simple.clinic.storage.text.LocalFileTextStore
-import org.simple.clinic.storage.text.TextStore
+import org.simple.clinic.storage.text.TextStoreModule
 import org.simple.clinic.user.User
-import javax.inject.Named
 
 @Module(includes = [
   RoomMigrationsModule::class,
   SharedPreferencesModule::class,
-  FileStorageModule::class
+  FileStorageModule::class,
+  TextStoreModule::class
 ])
 class StorageModule {
 
@@ -42,20 +40,5 @@ class StorageModule {
   @Provides
   fun userDao(appDatabase: AppDatabase): User.RoomDao {
     return appDatabase.userDao()
-  }
-
-  @Provides
-  fun textStore(
-      fileStorage: FileStorage,
-      @Named("reports_file_path") reportsFilePath: String,
-      @Named("help_file_path") helpFilePath: String
-  ): TextStore {
-    return LocalFileTextStore(
-        fileStorage = fileStorage,
-        keysToFilePath = mapOf(
-            "reports" to reportsFilePath,
-            "help" to helpFilePath
-        )
-    )
   }
 }
